make recursive update continue for constant no. of unique elements
sometimes, the number of unique connectivity values temporarily stagnates,
followed by an increase again.

To maximize the number of unique connectivity values in a molecule, we
incorporate a trial_number argument in the recursive 'update' function.

Also, we change the type of 'Vertex.connectivity' from short to int,
because of the large values that they become.