Added Exception for user input of double bonded O2
Many users erroneously use the double bonded singlet version for O2, which
does not react with most reaction families. Now RMG will throw an error
giving an explaination for why the user should use triplet O2.

If the user actually wants singlet O2, I have included a new option
"allowSingletO2" to the generatedSpeciesConstraints. Note that this
constraint is a little more restrictive by design. It will ignore
the allowed list, which usually lets users circumvent species
constraints for inputs such starting species and reaction libraries. The
only way to have singlet O2 is to set allowSingletO2=True. The point is
that we never want singlet O2 to fly under the radar.

Moved check for O2 singlet to just for initial Species

As suggested by rwest, it is more efficient to only check for singlet O2
when initializing the mechanism instead of for every species added. Right
now it checks every species from the input file and all species in seed
mechanisms and reaction libraries