Attempt to print non-aromatic SMILES of aromatic molecules.
eg. it would be neat if we could print SMILES of these three
molecules separately: 'CC1=CC=CC=C1O', 'CC1C=CC=CC=1O', 'Cc1ccccc1O'.
They all have unique adjacency lists in RMG terms, and so are
all different "Molecules" even if resonance isomers of the same Species.

As SMILES strings are used as shorthand to represent Molecule objects,
it would be nice to print the SMILEs strings differently.
I think this works.

I now think the remaining problem is not that RDKit can't create
the correct SMILES, it's that we can't read the SMILES correctly.
i.e. the round-trip test fails, but it's not creating the SMILES from
the RMG molecule, but creating the RMG molecule from the SMILES.
Problem is probably that RDKit reading from smiles automatically perceives
the aromaticity, then when we tell it to Kekulize, it does it its own (random)
way not the way that was specified in the original SMILES.