Modify documentation to remove some code cut offs
There isn't a way to wrap code in sphinx (https://github.com/rst2pdf/rst2pdf/issues/465),
so things get cut off our latexpdf documentation. I've tried to address it where possible, and
these should still fit on the page and not cause any problems. This addresses the issues present
in the RMG User guide, but there are others in the developers guide.