Change the way QMsettings sets symmetry executable path.
Either use the one found in RMG-Py/bin/symmetry if it exists,
or assume it's on the path.

Place a try except block when running the subprocess that
calls symmetry.  Raise an exception with a msg that tells the user
to install symmetry if it fails in a way that shows RMG cannot
find the executable.

This is to make both the source installation and conda build
(where symmetry is installed as a package) compatible
with RMG.