Add version number to __init__.py
This is so rmgpy.__version__ can be queried directly

Manually enter version numbers in both setup.py
and __init__.py.  There are ways to save it only
in one place and have it be read in or imported,
but since we also use version numbers in the conda
meta.yaml file, it wouldn't be easy to bypass manually
inputting it.  So I'm using the manual system for now.