Remove the CanTherm User guide as a separate pdf.
Having CanTherm guide and theory repeated within the RMG
user guide was a bit redundent.  Combining the two
makes sense in that the entire document is 100 or so pages.

I've changed the title to reflect that both types
of documentation are included.

Now there is distinctly 1 documentation pdf and 1 api reference pdf.