Renamed .isSurfaceMolecule() and .isSurfaceSpecies() to .containsSurfaceSite()
Not as nice a name, but at least it is the same for both
Species and Molecules, which means we can test both without
knowing or caring what they are.

If you think of a better name for this method, please change it!

Now the reactionTest.py unit tests pass