Incomplete work towards getting surface thermo data
Need to make a gas phase copy of a surface bound molecule
in order to estimate its thermochemistry. Unresolved issue is
whether to assume singlet or triplet (eg. lone pair or biradical)
for species that are double-bonded to the surface.
Perhaps need to do both and find the lowest energy.