Exception handling when trying to add species to core.
We should fix the bugs in RMG-Py or RMG-database that
cause these errors, but in the mean time this should allow
the importer to function by just skipping the problematic
reactions.

I made it as specific as easily possible, so only species that
cause a problem are skipped. (But we do however skip all reactions
of such species).