Allow # in chemkin species labels.
There are examples of this in the literature.

We could be more explicit and ensure they follow chemkin rules:

"Species names are composed of up to 16-character upper- or lower- case symbols. The names cannot begin with the characters +, =, or a number; an ionic species name may end with one or more +'s or -'s."