Force longDesc and shortDesc to be unicode
This will raise an error when you try to set it with
something that can't be encoded, rather than causing
problems downstream. It may slow things down a bit;
I'm not sure how much this matters.

This *might* not be helpful.

Also fixes unicode errors saving thermo library entry comments.