create hash and equality comparison methods for Species objects
the hash uses the fingerprints of the molecule objects

the equality operator uses the equality comparison of the molecule objects

a species A is equal to a species B if and only if the set of molecules
representing both species are equal.

a compare method allows for the comparison of a species to a molecule
object.