Conditionally set distGeomAttempts
We typically set it to 15 * (# atoms - 3), no real reason, but to give
a large number of embeds and relate it to the number of atoms. For certain
cases, such as [H][H]+[O]_[H]+[OH], we have 3 atoms, so the number gets set
to 0.

This is a rare case, as often in RMG we would use a seed mech or reaction
libraries for small molecules, but for the rare case where we might want to
do these calcs on small molecules, this was addressed.