Extend generateReverseRateCoefficient for some PDep types.
I think that the Troe broadening parameters etc. are 
the same in both directions, so as long as we correctly
reverse the high and low pressure limits, we should be 
able to reverse these reaction types too.

I think ThirdBody with +M and the low-pressure limits of
the PDep reaction types with (+M) should have one more
m3/mol in the units than they would otherwise.

I tried to reduce code duplication by making a special
method to reverse a given Arrhenius rate expression.