send species indices from workers to root
should reduce the amount of data transferred in the parallel scheme.

A dictionary with a unique index pointing to the core+edge species objects
is created as an attribute of CERM.

Every time a new species is created in makeNewSpecies, a new entry is
created in this dictionary.

When a reaction is returned from a worker, it needs to be "inflated",
i.e. the species represented by an index are converted to the corresponding
species object.