Add prioritizeThermo function
Always sorting by enthalpy may be problematic, because sometimes it
 actually uses a more generic group.
 Create this function to treat cyclic groups
in a special way so as to find the best match,
rather than the most stable (often wrong) correction.

Use the summed rank as a way to prioritize thermo. Do this only for rings right now.

First sorts by rank (ascending order), then by most stable enthalpy at 298K.
This only affects rings currently, which has a rank of 1 for aromatic ring corrections
and rank of 3 for all unlabeled groups, and rank of 10 for generic top nodes.