Add a CH3 adsorption test to rmgpy/solver/surfaceTest.py
The parameters were generated by an RMG-cat simulation.
If we replace the StickingCoefficient expression with a
SurfaceArrhenius expression (currently commented out)
then it works. As it is, it crashes the ODE solver.
Conclusion: the StickingCoefficient code is broken.

Fix the surface reactor solver unit tests, with nSims field.
Needed an nSimsTerm (later nSims) field
(although maybe a default should be OK?)