create new function areIdenticalSpeciesReferences
CERM.checkForExistingReaction

The current state only compares reaction labels, which are a
concatenation of species labels. There's a risk that two distinct
species may lead to the same label, hence resulting in a false
positive match in CERM.checkForExistingReaction.

The function areIdenticalSpeciesReferences check references
of reactants/products in both directions.