check maximumIsotopicAtoms constraint in constraints module
only count the number of non-normal isotope atoms in a molecule
when the 'maximumIsotopicAtoms' key is not equal to the default
value of -1.

Iterates over the atoms of the molecule object and compares the mass
of the current atom to the mass of the atom that has the same
symbol.