Located degeneracy bug for dissociative adsorption.
Still need to fix it, but at least I've found it.
Trouble is RMG previously assumed all elementary reactions are 
at most bimolecular in both directions. The reverse of 
dissociative adsorption is:
 AX + BX -> AB + X + X
with three products. We generated both reactions
 AX + BX -> AB + X + X
 BX + AX -> AB + X + X
but didn't recognize them as duplicates because we hadn't 
implemented a way to compare a list of three products.