More progress towards "termolecular" A + X + X -> reactions
Think it may even be working.
Stepped through quite far with methane adsorbing.
It finds 8 ways for methane to do dissociative adsorption:
R1 could be C and R2 each of the four H's, or the other way around.
Then it collapses this to one reaction with degeneracy 8.

There's a bit of code duplication that perhaps could be 
refactored away.