checks whether the solvent is listed as species
The solvent now has a new attribute called molecule. And all the species
has a new attricute called solventStructure and isSolvent. For
initialspecies, the code first checks whether the species is the solvent or
not by performing isIsomorphic test. isSolvent is True if the species is
the solvent and is False if not. By performing is Isomorphic test, RMG can
recognize that the certain species is the solvent although the string
names of that species and the solvent are different in the input file.
Then in the rmgpy/rmg/main.py, if none of the initial species is the
solvent, it raises Exception with an error message that "One of the
initial species must be a solvent"