prevents RMG from crashing when old databases used
because I added new solvent attribute 'molecule' that stores the solvent's
molecular structure, it requires the
RMG-database\input\solvation\libraries\solvent.py to be updated to the
newest verison that contains solvent's molecular structures in either
SMILES or adjacency list.
This new commit prevents the RMG from crashing when old RMG-database
without solvenet's molecular structures is used.
If the old RMG-database is used, liquidReactor will simply use the
solvent's string name to determine whether the species is the solvent or
solute.