added unittest for checkSolventinInitialSpecies
I added a new function in rmgpy.data.solvation.py called
checkSolventinInitialSpecies. This checks whether the solvent is listed as
one of the initialSpecies by looking at the molecular structures (when the
solvent SMILES is known) or by looking at the string name (when the
solvent SMILES is not provided). It raises Exception when the solvent
cannot be found in the initialSpecies. In rmgpy/rmg/main.py, this function
is used when the liquid reactor is used.