move deep copy of reactants from reactAll to react function
a deep copy of reactants is necessary to avoid modifying
reference objects that will be re-used in the next round of
reaction generation.

The move to react from reactAll will ensure that the deep copy
is taken on a slave worker instead of the root in case multiple
workers are available.