Change adjlist error message to show entire adjlist
Previously it assumed the first line of the adjlist was a label, and gave the first line. For adjlists without a label as the first line, this was not very helpful as an error message. I have changed it so that now it prints the whole adjlist so that it is useful in both cases and likely shows the error so that user can debug faster.