add more accurate solv thermo correction &unittest
Previously, the linear temperature dependence of the solvation free energy
was used for solvation thermo correction. Now with this change, a more
accurate thermo correction can be applied by using the linearity of
Tln(K-factor) w.r.t. (rho-rho_c) near the critical point of the solvent,
and the critical limit of the K-factor. Near 298 K, the Abraham LSER and
the Mintz LSER are used. Tln(K-factor) vs.(rho-rho_c) is expressed as
a 2nd order polynomial for 298 K < T < 380 K
and it switches to a linear function for 380 K < T < T_c.
The curve is smooth and continuous at 380 K.
The details of the new def are explained in rmgpy/data/solvation.py

The using this K-factor curve, the solvation free energy is calculated at
the specified reaction temperature. The entropy correction is set to the
constant value as the entropy of solvation at 298 K. The enthalpy of
solvation is then calculated so that (dGsolv = dH - T * dS) is satisfied
at the specified temperature.