Add kinetics comments when the rate rule is from a training rxn
Can be viewed in chemkin if verboseComments is set to True
Previously it would be very difficult to trace back where
the rate rule numbers are coming from unless we look at
live entries in the rate rule database.  The trained rate rules
are not hardcoded and generated on the fly at the beginning of the job.