Change kinetics average commenting, and don't average the kineticsList when there's just one entry
Deep copy the kinetics even if one entry is used
in the average, because we are going to be modifying the comments
and we don't want to modify the original object.

Use just the children's template labels for an averaged rule's comments.
This modifies the commenting that we used to do, which
traces the entire averaging history by recursively
appending to the kinetics comments in order to
view the actual original templates that factor into
the average.  Instead this new method of writing
the kinetics comments just lists the distance 1 children that
in fact has kinetics.  In order to do more detailed analysis, one
would have to go inside to uncomment the debug lines
that use the former kinetics comment style that retains
full information.  However, this is too memory intensive
and difficult to read for normal RMG jobs, and wil print
out too much to verbose chemkin output using the original method.