Possible fix for the duplicate training reactions problem
Bug was found where reaction 2 in  R_Addition_MultipleBond/training
C2H3O3 <=> C2H2O + HO2
was getting duplicated in the final model.  It appears
that we fail the checkForExistingReaction check because we typically
only check in reverse direction when the family is its ownReverse.
However, this forgets that there are some training reactions that
are in fact in the reverse direction!

Therefore we should more expand the searchRetrieveReactions()
function to generate reverse direction reactions for all families.
This may slow things down, leave optimization till later.