Finished(?) updating the filterReactions script.
This tool loads all the kinetics library files that can be 
found in the path given in the command line argument.
It finds reactions that are XH + .OOH <=> X. + H2O2
and saves them in a big dictionary of dictionaries
in a pickle file and python file both called kineticsDict.