Unit test for lone pair/biradical thermo.
If we're estimating via group additivity, a species with lone pair
or biradical on the same atom should give the same thermo result.
Testing on a cyclopropane birad C1C[C]1.
Otherwise, we fall up to more general, erroneous groups since the
lone pair form is not a radical, and it's not bonded to hydrogens.