Time integrated flux can be chosen for analysis
Users can choose either instantaneous flux (mole/cm3/s) or time
integrated flux (mole) for flux analysys. The time integrated flux is
integrated up to the investigated time point.

Also, units for rate coefficients, instantanoues flux, and integrated flux
are all displayed.

The arrow direction for the product formation pathway is switched, so that
the direction of the reaction and the sign of the flux matches