adds a more deterministic method for finding sssr
so that bicyclic decompostion is more deterministic when dealing with
molecules like C1CC2C3CSC(CO3)C2C1. The basic idea behind is to break
ties when selecting rootVertex or cycle by adding more sorting keys.
The newly added keys in this patch are atomic numbers of the atoms along
the rings and original connectivity values.