checkSolventStructure method is added
In rmgpy/data/solvation, the checkSolventInitialSpecies method is
replaced by checkSolventStructure method. It takes the solvent
class as the input, loads the corresponding solvent structure from the
solvent database by using the getSolventStructure method, and checks
whether the solvent has the correct molecular structure by performing
isIsomorphic test. If the solvent's molecular structure is different from
the one listed in solvent database, it raises DatabaseError.
If the old version of the database is used, the loaded solventStructure is
None, so this test will be skipped.
The corresponding unittest is also added in rmgpy/data/solvationTest.py
The checkSolventStructure method is used in rmgpy/rmg/main when it
performs all liquid-phase startup things. solventData is also loaded in
rmgpy/rmg/main.
Checking whether the solvent is listed as onoe of the initial species is
done in rmgpy/rmg/input.py (previous commit)