add a function that fits NASA to free energy data
Given the array of temperatures and free energy data, a new function
fitToFreeEnergyData can fit the NASA model to these and return a new NASA
polynomial. This will be used to fit the species solvated/liquid phase
thermo after the solvation free energy correction is made.

The corresponding unittest is made in rmgpy/thermo/nasaTest.py. The fitted
enthalpy, entropy, and free energy agree with the original values within
reasonable accuracy. Because the heat capacity is obtained by
differentiating the free energy twice, the fitted heat capacity has higher
error. However, as long as the fitted enthalpy, entropy, and free energy
are accurate enough, this method is fine because the heat capacity is not
actually used in liquid reactor.