Fix to gain/lose lone pair functions for GroupAtoms
Previously didn't seem to work unless self.lonePairs = None, which always seemed to be the case.
Now if there is self.lonePairs, we set it to a list equal to self.lonePairs + 1 for gain lonepair and self.lonePairs -1 for lose lonepair.
If self.lonePairs = None, we set it to a list equal to [1,2,3,4] for
gain lonepair and [0,1,2,3] for lose lonepair.