Change product template duplicate testing from isIsomorphic to isIdentical
Previously we used isIsomorphic to delete duplicate product templates, when we really should be using isIdentical which does not respect wildcards. This make it so that we could potentially lose product templates like R u1 if C u1 was added first.

This push may make RMG find many more reverse reactions for family's that are not their own reverse.