New methods to remove Van der Waals bonds from molecules or groups.
This current concept is that an adjacency list for a van der waals
bond does not actually store any bond, because you don't know which
atom it should be bonded to. You just have the surface site as
part of the same object, but not bonded to the rest of the structure.
However, after applying a reaction recipe, you will have an explicit
van der Waals bond to one of the labeled atoms. These methods allow
you to remove all such bonds.