separated reduction of degeneracy of sameSpecies from findDegeneracy
Since the generateReactions requires degeneracy calculation, so the
`reverse` attribute can be applied, two calls must be make to
findDegeneracies, one during generateReactions and another for
all of the resonance isomers in React. The calculateDegeneracy function
was modified to allow for degeneracy to be applied separately from
degeneracy calculation.

squash! separated reduction of degeneracy of sameSpecies from findDegeneracy

eventually, the reduction of degeneracy for same species can be removed
this squash removed calculateDegeneracy from the generateReactions method.