Always use scaled harmonic oscillator (HO) frequencies for calculating ZPE in Cantherm, rather than projected frequencies when hindered rotors are included.
Ideally, the HO and projected frequencies should not
vary mucy except for the removal of the mode corresponding to the
internal rotation in the latter case. However, in cases with strong
coupling between modes (e.g., hydrogen bonding) the projected
frequencies can be much different that the HO frequencies, resulting in
ZPE's that can differ by kcal's, which has  a proporitional effect on
thermo results, and potentially an exponential effect on kinetics.

Because the ZPE represents the internal energy from the lowest possible
vibrational modes at 0 K, it is reasonable to model all vibrations as
HO's at that condition, including modes corresponding to internal
rotation.