Improve handling of false positives during aromaticity detection
Previously, if any ring in the molecule was falsely identified as
aromatic by RDKit, then generateAromaticResonanceStructures would
completely give up on trying to make the aromatic form.

Now, the algorithm will go through each ring in the molecule that
has been identified as aromatic, and create a final molecule with
as many aromatic rings as possible, ignoring false positives.

Conflicts:
	rmgpy/molecule/resonance.py