Reorganize thermoTest.py
Previously, TestThermoDatabaseAromatics was rerunning all of the
tests in TestThermoDatabase even though not all of the tests were
written with that intention. Now, the tests originally designed for
that have been separated into a new class called TestThermoAccuracy.
ThermoDatabase tests that were intended to only be run once were
kept in TestThermoDatabase.

Minor updates were also made to docstrings and formatting.

Note that both tests in TestThermoAccuracy are still marked
@work_in_progress!