Adding details to docstrings about what some methods return.
This helped us figure out what was going on when picking our
way through the code. Eg. some bits of code make flow decisions
based on whether the third thing returned in a tuple is None
or not, so its helpful to know what that thing is expected to be.