Added option in getInternalReducedMomentOfInertia
Added options in getInternalReducedMomentOfInertia to calculate the reduced moment of inertia based on the bond axis and an axis parallel to the bond axis, but passing through the center of mass of the respective group.

updated the docstring

set default option to 3 as recommended in East, A. L. L. and Radom, L. *J. Chem. Phys.* **106**, p. 6655-6674 (1997).