make cantherm comment out excess reactions
this commit modifies cantherm to only output reactions that are independent.
It checks isomorphism with all the reactions that have been printed and
comments out those that are isomorphic, since their reverse reaction has been
printed.

This commit also refactors the cantherm.pdep save method so that saving
to 'output.py' and 'chem.inp' occurs in the same loop to reduce repetitive
code.