Fix bug where some species did not have pictures saved in diffModels tool
For the options diffOnly and commonDiffOnly, we only saved species images if they had different thermo and reactions if they have different kinetics. However, its possible that some of these differing reactions contain species that have exactly the same thermo. In these cases, the reaction would not display properly because we never saved some of the species.

This commit fixes this bug