Reading\writing Chemkin files with a specificCollider
Generalizes reading the third body collider whether it is (+M), (+m) or
a specific species, i.e., (+N2).
Gives RMG the ability to read the collider even if the species name has
paranthesis, e.g., N2(5) or CH2(T)(15).
This is still not implemented in Chemkin and Cantera, but both have open
issues to implement it in the future.
Also adds a comment to verbose Chemkin file that a specific third body
collider was identified.