expand rxn.isIsomorphic to optionally check species labels
when RMG isn't able to tell the difference between two species,
like enantiomers, but they would like to be kept distinct, the
checkOnlyLabel option is added to rxn.isIsomorphic which checks
species' labels instead of isomorphism.

if it becomes useful to check label and isomorphism, either the user
can call `isIsomorphic` twice or refactor the code to check both
label and isomorphism.