Adding an Exception to retrieveNewSpecies()
The retrieveNewSpecies(self, deflatedRxn) function assumes each
deflatedRxn sent to it has at least one reactant/product which is an
integer (i.e., in the core). This is a very logical assumption if
everything works as expected, but we should throw an error if this `for`
loop terminates without returning a thing. see #1078