Changes to fixBarrierHeight
Should fix #1086

Now ought to address the problem of exothermic reactions
giving negative activation energies as a result of Evans-Polanyi
expressions for any value of E0  (previously it only worked if E0>0).

The aim of this code is as follows:
Calculated Ea values are allowed to be as negative as E0 if E0
is negative, but if E0 is positive then Ea cannot be below 0.