Read externalSymmetry and spinMultiplicity from QM job output
Read externalSymmetry and spinMultiplicity from QM job output if
unspecified in CanTherm input.
Previously an error was thrown if either externalSymmetry or
spinMultiplicity were unspecified in CanTherm input, although we are
able to read both parameters from gaussian/QChem outputs.
Since externalSymmetry is often incorrect (at least in QChem), a general
warning is given if this value is unspecified