moved RMG exception class to specific file
When importing an RMG exception, before this commit, all the
imported modules for that file are automatically imported, which
can sometimes lead to circular dependency issues (see #1101).

This commit centralizes all exceptions within RMG into rmgpy.exceptions
and does not require any imports, to avoid any circular dependency
issues. This also should make it easier for people to find and use
more specific exceptions in RMG in the future.