add option to vary accuracy and speed of findDegeneracy
When accuracy is prefered, this method uses a robust O(n2) algorithm
to ensure the correct degeneracy is output. When speed is prefered,
this method uses a faster O(n) algorithm to both remove redundant
reactions and approximate degeneracy, which is similar to the previous
degeneracy algorithms. Only resonance structures are generated for
the more robust algorithm.

Also, a previously unused option in findDegeneracy was removed.