get proper degeneracy for rxns added to the core
When a reaction is added to the core, RMG will compute the accurate
degeneracy of the reaction using family.calculateDegeneracy.

This change broke some unittests for modelTest since using
calculateDegeneracy in model.py requires real examples of
which degeneracy can actually be calculated. So the unittests were
replaced with realistic examples.