always initialize reaction diffusion attribute
This causes k_effective_cache to always be initialized to an empty dictionary rather than not being initialized if the diffusionLimiter is enabled.  The issue with leaving it empty is that if you initialize diffusionLimiter after the reaction is created rate calculation breaks by raising an error that is caught in a way that cause rates calculations to always return 0 which results in distant errors down the line.  Working around this in any context except a standard RMG run can be very inconvenient.