Remove usage of Species.solventData, solventStructure, solventName, solventViscosity
Used the getDB method to avoid needing to store solventData within
the Species class.

rmg/main and tools/sensitivity, removed setting of the attributes

solventStructure involves storing a species object within a species
object. This seems like a bad idea. This commit modifies other parts
of RMG so the solventStructure attribute is no longer used, since
the functionality does not need to be within species objects. Future
commits remove this attribute.

The solventName should not be an attribute of each species. This
method passes solventName through thermoEngine to avoid storing
the solventName string in species.