Simplify model.reactionDict
model.reactionDict was not very intuitive to use since it was 3
layered dictionaries with a list underneath. This commit simplifies
the reactionDict by making it a single dictionary with the key and
value being the same reaction object. This currently could not be
made into a set because checkExistingReaction requires the current
reaction to be output, which isn't possible for a set. This commit
also removed the numerous helper methods for the previously
complex reactionDict object.

This should improve readability and usability of the model.py
module.