change kinetics commenting of training reactions
Previously kinetics comments from training reactions started
with the rate rule. Due to the way loadChemkinFile parses the
comments by identifying starts of lines, this caused parsing errors.
This commit reworded that line of the comments to start with
identifiable words. It also updated the uncertainty module parsing
to work with this change as well as updated the unittests.