automatically load the library a reaction is from
since RMG requires a library to be loaded for it to load a reaction from that library, reactions in a (created seed) library that are from a different reaction library need their original library to be loaded, adding these libraries as seeds or reaction libraries risks double counting or errors and isn’t terribly intuitive anyway. This commit automatically loads these libraries when their first reaction shows up without adding them to edge or core.