Ensure that superimposed species connect to main flux graph
Previously, adding central species to a flux diagram did not take into
account connections to the main graph. Now, specifying the superimpose
keyword ensures that such connections are made in a recursive
algorithm. At each step, the maximum flux connections (with the number
of connections determined by centralReactionCount) are added until all
central species connect to the main flux graph. If superimpose is
used, rad is set to -1.