Revert "added in declarations for Pmax and Pmin in arrhenius.pyx"
This reverts commit cb45c2f7afa66b4e1cdc014eaa6ff1db140dc8e3.

The Pmax and Pmin should be set on the Arrhenius object by the
call to KineticsModel.__init__, so I don't think we should have
to wipe it and set them to None. Perhaps there was a problem
with the testing database at the time this commit was made.
I'm going to revert the commit and see what happens. If it causes
a problem, I'll look for a[nother] solution.