Allow addLabelsForReaction to properly handle identical references
If a reaction object, which had Molecule or Species objects with
identical references, was input to addLabelsForReaction, the
conflicting references would not allow the same molecule to hold
multiple labels, leading to errors.

This creates a deep copy of the Species object when identical
references are found, since seprate references are necessary for
the labeling.

This commit also adds three test methods for addLabelsForReaction
which previously had no test methods.