Enable trimolecular reaction families
When generating reactions, check for identical reactants for
degeneracy calculations, and implement isomorphism checking for
trimolecular reactions. Generate reactions from all relevant
permutations of the three reactants. Also allow rate constant units
for trimolecular reactions. Use isomorphic_species_lists for
isomorphism checking. Also, reduce code duplication in
getLabeledReactantsAndProducts.

Additionally, reverse the reactant order for bimolecular reactions
when generating reactions with swapped reactants. Previously, when
checking reaction generation from swapped reactions in bimolecular
families, only the reactant template order was swapped, but not the
order of reactants. For families with the same template for both
reactants, this means that not all products can be generated if they
depend on the order of reactants. The dependence on the reactant order
originates from the relabeling of atoms that happens when applying the
recipe.