Allow trimolecular reactions in RMG simulations
Add trimolecularReact and trimolecularThreshold arrays to enable
reactions and reaction filtering of trimolecular reactions. To not
do unnecessary computations and require unnecessary memory, use a
trimolecular flag in RMG and ReactionSystem which is set to True in
RMG.loadDatabase and RMG.initialize if families with three reactants
or products are found.

Also add trimolecular reactions to generate_reactions utility.