creation of getExtensionEdge function
This function finds the edge of the extensions to a given node (the smallest set of extensions such that any extension matching the original group and at least one reaction that matches that group must be subgraph isomorphic to a member of this set of extensions).  Speed relies heavily on identifying dimensions that don't affect optimization and labelling them so that they are ignored.

Regularization list attributes are filled with all the values of that attribute that occur in training reactions matching that node