Don't generate resonance isomers for aug-InChI
As clearly stated in the InChI paper:
"Mesomerism is effectively eliminated in InChI. Mesomers have the same
InChIs (this is true for all possible InChI layouts of layers).
Actually, this is very natural. Mesomeric structures of a molecular
entity have the same basic connectivity but differ in bond orders, and
maybe by having atomic charges on different atoms. InChI does not use
bond orders and does not place charges on particular atoms; the
placement of hydrogen atoms in a mesomeric system, which would be
important for InChI, is always the same."
Source: InChI, the IUPAC International Chemical Identifier,
J. Cheminform 2015, 7, 23, doi: 10.1186/s13321-015-0068-4