RMG memory and condition generation
this system allows RMG to choose new conditions for each simulation using the
Weighted Stochastic Grid Sampling Algorithm (idea credit to Kevin Silmore) discussed below.

First all conditions are normalized (P is normalized in log P units).  An objective function is defined that weighs conditions based on their distance from prior conditions, on the number of objects returned and how recent the prior condition was. This objective function is evaluated in a multidimensional grid and these evaluation points are normalized so that they sum to one.  Then a condition is chosen randomly with each point being chosen with the probability of its normalized objective function value.  After this a random step (maximum magnitude is sqrt(2)/2 times the distance to another grid point) is taken from the grid point to give a final condition.  This causes it to sample from the entire condition range.

The random number generator used is seeded so the process is deterministic.

The RMG_Memory object keeps track of all of this. The number of simulations completed successfully (without adding objects) is also tracked to determine when the reactor can complete.