Add a pruning counter to fix issues with ranged reactors and pruning
The pruning counter counts simulations up to max(nSimsTerm for each reactionSystem) * # of reaction systems (after this number of simulations the reactor with the largest nSimsTerm value can be assumed to have sampled the whole space).  When the counter is below this value pruning will not occur, once pruning happens the counter is reset at 0.