Handle invalid vector errors gracefully in clar optimization
For certain polycyclics, getAromaticRings can give inconsistent
results from call to call, as a result of some non-deterministic
behavior in RDKit. If the number of aromatic rings changes, then
the size of the matrix will change and previously generated
constraint vectors will be the wrong size. Since this does not
occur very often and occurs mostly with odd polycyclics, it seems
more efficient to catch this error and continue instead of ending
the RMG run.