Fix networkIndices for edge networks
When a high-p limit rate from a library reaction is converted into a
pressure dependent rate upon library loading, a network is created with
its source species in the edge. Since the speciesIndex dictionary only
contains core species, a KeyError is raised.
This fix captures such cases and assignes an index value of `0` to the
respective matrix entry. The only usage of networkIndices is in
residucal() under simple.pyx, where these values are compared to `-1` to
derive the number of reactants.