Add option for mol.getFormula to separate isotopes
getFormula is used for many different checking operations, and
some of these should be descriminating between isotopologues.

This commit adds an option to getFormula to allow separate isotopes
to be added separately. A unittest is created, and this option is
applied to when getFormula is used as a shorcut to certain SMILES.