Fix Uncertainty after RangedReactor refactoring
The Local uncertainty ipynb was not able to run properly after
the nSims parameter was added to the middle of the arguments when
initiating SimpleReactor, since the arguments sent to the
SimpleReactor from the uncertainty module were all shifted.

This commit/PR fixes this by using keyword arguments in the
uncertainty module.