TEMPORARY - changes to databaseTest to help debugging.
It now skips loading (and hence checking) all the kinetics libraries
and families, and it also prints a bunch of debug information when
checking the thermo group entries are accessible.

This was to track down what was causing a "Bus error: 10" for me
as mentioned in https://github.com/cfgoldsmith/RMG-Py/issues/42.

This commit should not end up in Master, when merging.
(Or at least it should be reverted).