Allow hasStatMech to assess atomic molecules
Previously, hasStatMech would return True for any molecule with
one atom, even if the E0 parameter was None. This raises issues
when creating pdep networks since E0 is necessary. At the same
time, this method does not check E0 parameters even for polyatomic
molecules.

This commit reworks hasStatMech to differentiate between atomic
molecules, which it only checks the E0 parameter, and polyatomic
molecules, which it checks both E0 and the number of modes.