Allow pressure-dependent gas-phase kinetics in surface reactor (with error).
This will mean the simulations are (at least slightly) wrong, but may
allow things to not crash.
The trouble is in generate_rate_coefficients(self, coreReactions, edgeReactions)
we pre-calculate a bunch of k values, which are assumed to stay constant
during the ODE solve, so the ode solver just multiplies them by the species
concentrations (with relevant stoichiometry powers) to get the rates of progress.

We have never had to re-calculate the k values during each call to the
right hand side of the ODE equation before, because we always had
isothermal, isobaric, conditions.

This tries to hide issue https://github.com/cfgoldsmith/RMG-Py/issues/54