Chemkin writing uses Units.getConversionFactorFromSItoCM()
This will allow us to do surface reactions with correct units.
(Which will be implemented in a later commit)

For gas-phase reactions, I put in a bunch of assert statements
to check that the new implementation has the same effect as 
the old.

If this is speed-critical code, they could be deprecated/removed.