Fix rate constant unit conversion to cm/mol/s
Replaced the method getConversionFactorFromSItoCM
which used to leave things like Molecules in their 
original form, with a new method
getConversionFactorFromSItoCmMolS
which converts everything into S (apart from length
becomes cm). This means that all chemkin files are
now written in CM/MOL/S combinations even if the original 
rate was in m3/molecules/minute.

A few reaction libraries and training rules are stored
in molecules instead of moles, but the chemkin file
must be consistent.