Get Arkane to use Symmetry package for symmetry/optical isomers
Created new method get_optical_isomers_and_symmetry_number in Log class
This method allows Log objects to calculate the number of optical
isomers and symmetry numbers contained within a log file.
This was taken and modified from the equivalent method in ARC (credits to Alon).

Remove symfromlog parameter. If using the RMG QM symmetry package,
the symfromlog parameter is not necessary and confusing.
This commit removes the need for such a parameter.

Get Arkane to use Symmetry package for symmetry.
This commit uses the Symmetry package to estimate symmetry if not
explicitly given by the user by modifying the gaussian, molpro,
and qchem parsers.