Set ArkaneSpecies properties after loading the objects
Some properties are important to parse, but are not objects and cannot
be updated automatically.
`imaginary_frequency` contains a ScalarQuantity instance, but it is not
an object in itself in RMG/Arkane, just an attribute of TransitionState
(.frequency). It's important to parse it for tunneling.
Structure information is important to parse since YAML file generation
depends on it.