ApplyRecipe returns reaction adjacency lists if requested, for reporting.
If requested, applyRecipe now returns the reaction adjacency lists
as as a tuple.
If expected to return a tuple, do so even when the result is None
But don't add adjacencyList to rxn if rxn is None.
Be careful where you generate the reactant adjacencyList, before manipulating it.

Purpose: to aid reporting of these for third-party tools to search for transiton
states automatically.