Fix bug in KineticsModel.isIdenticalTo
This method just checks that Tmin and Tmax are equal.
But if they were equal, it used to return False.

Now it returns False if one or other is None, or they're
different. Note, however, that because the Quantity.equals
method defaults to a 1% tolerance, that's 20K difference
needed at 2000K.