Move PyDAS solver check from Makefile to utilities.py
Primary goal is to not overwrite settings.pxi if it already exists.

When using `cythonize`, Cython does a better job of figuring out
if files have changed and need to be recompiled. Since we would
rewrite settings.pxi every time upon calling `make solver`, Cython
would think that it needs to recompile the entire solver module
since it depends on that file. This change prevents excessive
rewriting and recompilation.