Fix normalizing of density of states
If a network has molecules with high energy, then normalizing by
exp(-Elist/RT) could lead to python to round to zero, creating NaN
in the density of states.

This commit checks for this numerical error and throws an
error if it occurs, so that the user knows the source of the mistake.
This error is most likely to occur if atom corrections are not used
or incorrect.

This commit also normalizes the product density of states, so
that all are consistant, unless there are no density of states
at any grain, which in that case no normalization occurs to prevent
divide by zero errors.