Fix normalizing of density of states
If a network has molecules with high energy, then normalizing by
exp(-Elist/RT) could lead to python to round to zero, creating NaN
in the density of states.

This commit fixes this problem by changing the exponential to
exp(-(Elist-Elist.min())/RT), which reduces any large values and
allows the normalization to not have numerical issues with high
energy values.