Don't error if utilities.py is missing in rmg.py
We check that the user is using python 3 calling code in utilities.py
This file is not present though in the binary install. Thankfully, this
check is also not necessary in the binary install, as conda ensures the
correct python version for us. Therefore, we don't need to throw an error.