fixed a bug with solvation GAV not finding radical
The solvation group additivty method had a bug that it cannot apply
radical correction to the radical species even if its radical group
exists in solvation radical database.
It was because originally, when it was applying correction for each
radical found, the for-loop range started from
range(1, number of radical), instead of starting from
range(0, number of radical). Because of that, if the number of radical
is 1, the correction was never applied.

This commits fixes this error by letting the range to be just
range(number of radical) since the number of radical is always integer