Improve Emax determination when solving pdep networks
Previously, Emax was set to a fixed value above the highest
energy isomer. However, for large molecules, the hard-coded value
may not be sufficient to capture the full equilibrium distribution.

The result was that Arkane/RMG would not be able to successfully
evaluate rate coefficients in these networks.

This commit implements a more complex approach of evaluating
the equilibrium distribution and then adjusting Emax until the
population at Emax is a specified fraction of the most populated
energy.