Differentiate CCSD(T) energy location with space
Previously, Arkane's reading of gaussian files would find an unexpected
string match in post run output (shown below), which Arkane cannot
parse. The proper line for reading constains a space, so the string
matching was changed to `CCSD(T)= ', which only picks up the intended
line.

Here is the line inadvertantly caught in the parsing of files:

 5327382\CCSD=-79.5337667\CCSD(T)=-79.5408947\RMSD=8.685e-10\Thermal=0.